Remove 'instabot/api/api_profile.py' because it allows for simplifications
