+++ MANIFEST
@@ -5,7 +5,6 @@ instabot/__init__.py
 instabot/api/__init__.py
 instabot/api/api.py
 instabot/api/api_photo.py
-instabot/api/api_profile.py
 instabot/api/api_search.py
 instabot/api/api_video.py
 instabot/api/config.py
